backend/requirements.txt
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
cryptography==44.0.3
python-dotenv==1.1.0
bcrypt==4.3.0
gunicorn==23.0.0
celery==5.4.0
redis==5.2.1
requests==2.32.3
pydantic==2.11.1
httpx==0.28.1
joblib==1.4.2
numpy==2.2.6
pandas==2.2.3
scikit-learn==1.6.1
xgboost==2.1.4
matplotlib==3.10.1
seaborn==0.13.2
pytest==8.3.5
